+++ tests/cpp/core/CommandLineOptionsTests.cpp
@@ -52,7 +52,6 @@ BOOST_AUTO_TEST_CASE(testCommandLineDefaults)
 
     BOOST_CHECK(!options.getHelp());
     BOOST_CHECK_EQUAL(options.getStreamId().toStdString(), "");
-    BOOST_CHECK_EQUAL(options.getPixelStreamerType(), PS_UNKNOWN);
     BOOST_CHECK_EQUAL(options.getUrl().toStdString(), "");
     BOOST_CHECK_EQUAL(options.getConfiguration().toStdString(), "");
 
@@ -65,7 +64,6 @@ BOOST_AUTO_TEST_CASE(testCommandLineDefaults)
 void setOptionParameters(CommandLineOptions& options)
 {
     options.setStreamId("MyStreamer");
-    options.setPixelStreamerType(PS_WEBKIT);
     options.setUrl("http://www.perdu.com");
     options.setHeight(640);
     options.setWidth(480);
@@ -76,7 +74,6 @@ void checkOptionParameters(const CommandLineOptions& options)
 {
     BOOST_CHECK(!options.getHelp());
     BOOST_CHECK_EQUAL(options.getStreamId().toStdString(), "MyStreamer");
-    BOOST_CHECK_EQUAL(options.getPixelStreamerType(), PS_WEBKIT);
     BOOST_CHECK_EQUAL(options.getUrl().toStdString(), "http://www.perdu.com");
     BOOST_CHECK_EQUAL(options.getHeight(), 640);
     BOOST_CHECK_EQUAL(options.getWidth(), 480);
@@ -84,8 +81,8 @@ void checkOptionParameters(const CommandLineOptions& options)
                       "/path/to/configuration.xml");
 
     BOOST_CHECK_EQUAL(options.getCommandLine().toStdString(),
-                      "--type webkit --width 480 --height 640 "
-                      "--streamid MyStreamer --url http://www.perdu.com "
+                      "--width 480 --height 640 --streamid MyStreamer "
+                      "--url http://www.perdu.com "
                       "--config /path/to/configuration.xml");
 }